+++ sketch_oct2a/MomentaryButton.cpp
@@ -4,7 +4,7 @@
 // When the InputPin is HIGHT -> the button is not pressed
 // Pressing the button shorts to GND and sends LOW to InputPin
 
-MomentaryButton::MomentaryButton(int p) : IODevice(p)
+MomentaryButton::MomentaryButton(uint8_t p) : IODevice(p)
 {
     pinMode(p, INPUT_PULLUP);
     Serial.print("Registered a button on pin ");
